# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HotelMate API
  version: 1.0.0
  description: Production-ready tenant-aware hotel operations, revenue reporting, approved-knowledge conversations, security audit trails, and live events. Every HTTP response includes an X-Request-ID correlation header.
servers:
  - url: /api/v1
security: []
paths:
...
  /guest/services:
    get:
      summary: List the authenticated hotel's active service catalog
      security: [{ guestBearer: [] }]
      responses:
        '200':
          description: Quick actions first, followed by the full active catalog
          content:
            application/json:
              schema:
                type: object
                required: [services]
                properties:
                  services: { type: array, items: { $ref: '#/components/schemas/Service' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /guest/requests:
    get:
      summary: List the current stay's service requests and persisted history
      security: [{ guestBearer: [] }]
      responses:
        '200':
          description: Stay-scoped requests
          content:
            application/json:
              schema:
                type: object
                required: [requests]
                properties:
                  requests: { type: array, items: { $ref: '#/components/schemas/ServiceRequest' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create a service request for an active stay or eligible pre-arrival upsell
      security: [{ guestBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serviceId, quantity]
              properties:
                serviceId: { type: string, format: uuid }
                quantity: { type: integer, minimum: 1, maximum: 20 }
                notes: { type: string, maxLength: 500 }
      responses:
        '201':
          description: New request with a persisted created event
          content:
            application/json:
              schema:
                type: object
                properties:
                  request: { $ref: '#/components/schemas/ServiceRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /guest/requests/{id}/cancel:
    post:
      summary: Cancel the current stay's request while it is still new
      security: [{ guestBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Cancelled request
          content:
            application/json:
              schema:
                type: object
                properties:
                  request: { $ref: '#/components/schemas/ServiceRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /staff/services:
    get:
      summary: List active and inactive hotel services
      security: [{ staffBearer: [] }]
      responses:
        '200':
          description: Tenant-scoped catalog
          content:
            application/json:
              schema:
                type: object
                required: [services]
                properties:
                  services: { type: array, items: { $ref: '#/components/schemas/Service' } }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Create a hotel service
      description: Restricted to hotel administrators and operations managers.
      security: [{ staffBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceInput' }
      responses:
        '201':
          description: Service created
          content:
            application/json:
              schema:
                type: object
                properties:
                  service: { $ref: '#/components/schemas/Service' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /staff/services/{id}:
    patch:
      summary: Update, reorder, price, or activate a hotel service
      description: Restricted to hotel administrators and operations managers.
      security: [{ staffBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceInput' }
      responses:
        '200':
          description: Updated service
          content:
            application/json:
              schema:
                type: object
                properties:
                  service: { $ref: '#/components/schemas/Service' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
  /staff/requests:
    get:
      summary: Read the role-aware operational request queue
      description: Housekeeping and F&B accounts are always constrained to services assigned to their fulfillment role; category remains an optional filter inside that role-scoped queue.
      security: [{ staffBearer: [] }]
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/ServiceRequestStatus' }
        - in: query
          name: category
          schema: { $ref: '#/components/schemas/ServiceCategory' }
        - in: query
          name: assignedToMe
          schema: { type: boolean }
        - in: query
          name: unassigned
          schema: { type: boolean }
      responses:
        '200':
          description: Priority-ordered tenant and role-scoped queue
          content:
            application/json:
              schema:
                type: object
                required: [requests]
                properties:
                  requests: { type: array, items: { $ref: '#/components/schemas/StaffServiceRequest' } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /staff/requests/{id}/assign:
    post:
      summary: Assign a request to an active hotel staff account
      security: [{ staffBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                staffId: { type: [string, 'null'], format: uuid, description: Defaults to the current staff member. }
      responses:
        '200':
          description: Assigned request with an appended history event
          content:
            application/json:
              schema:
                type: object
                properties:
                  request: { $ref: '#/components/schemas/StaffServiceRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /staff/requests/{id}/priority:
    post:
      summary: Change operational priority
      security: [{ staffBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [priority]
              properties:
                priority: { type: integer, minimum: 0, maximum: 3 }
      responses:
        '200':
          description: Prioritized request
          content:
            application/json:
              schema:
                type: object
                properties:
                  request: { $ref: '#/components/schemas/StaffServiceRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /staff/requests/{id}/transition:
    post:
      summary: Move a request to in-progress, completed, or cancelled
      security: [{ staffBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [in_progress, completed, cancelled] }
                note: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Updated request with an appended status event
          content:
            application/json:
              schema:
                type: object
                properties:
                  request: { $ref: '#/components/schemas/StaffServiceRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /staff/requests/{id}/notes:
    post:
      summary: Append an operational note to request history
      security: [{ staffBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                note: { type: string, minLength: 1, maxLength: 500 }
      responses:
        '200':
          description: Request with appended note
          content:
            application/json:
              schema:
                type: object
                properties:
                  request: { $ref: '#/components/schemas/StaffServiceRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
...
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    staffBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT with the hotelmate:staff audience.
    guestBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT with the hotelmate:guest audience and an active stay.
  schemas:
...
    ServiceCategory:
      type: string
      enum: [housekeeping, food_beverage, transport, wellness, other]
    Service:
      type: object
      required: [id, code, name, description, category, icon, fulfillmentRole, estimatedMinutes, priceCents, currency, isPaid, isQuickAction, isPreArrival, availableFrom, availableUntil, sortOrder, isActive]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        name: { type: string }
        description: { type: string }
        category: { $ref: '#/components/schemas/ServiceCategory' }
        icon: { type: string }
        fulfillmentRole: { type: string, enum: [reception, housekeeping, food_beverage] }
        estimatedMinutes: { type: integer, minimum: 1, maximum: 1440 }
        priceCents: { type: integer, format: int64, minimum: 0 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        isPaid: { type: boolean }
        isQuickAction: { type: boolean }
        isPreArrival: { type: boolean, description: Allows this paid service to be ordered before hotel arrival. }
        availableFrom: { type: string, pattern: '^(|([01][0-9]|2[0-3]):[0-5][0-9])$', description: Optional local hotel-timezone start time. }
        availableUntil: { type: string, pattern: '^(|([01][0-9]|2[0-3]):[0-5][0-9])$', description: Optional local hotel-timezone end time; earlier values create an overnight window. }
        sortOrder: { type: integer, minimum: 0 }
        isActive: { type: boolean }
...
    ServiceRequestStatus:
      type: string
      enum: [new, in_progress, completed, cancelled]
...
    ServiceRequest:
      type: object
      required: [id, status, priority, quantity, notes, totalPriceCents, service, events, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/ServiceRequestStatus' }
        priority: { type: integer, minimum: 0, maximum: 3 }
        quantity: { type: integer }
        notes: { type: string }
        totalPriceCents: { type: integer, format: int64 }
        service: { $ref: '#/components/schemas/Service' }
        assignedTo:
          oneOf:
            - $ref: '#/components/schemas/Staff'
            - type: 'null'
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        cancelledAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        events: { type: array, items: { $ref: '#/components/schemas/ServiceRequestEvent' } }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Role is not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Unique resource already exists
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: Input validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...